requirements.txt
django==3.2.0
django-mptt==0.14.0